First steps towards transferring SMC into ask/tell
