+++ pints/_sequential/_sequentialMC.py
@@ -26,127 +26,8 @@ class SMC(pints.SMCSampler):
     def __init__(self, log_posterior, x0, sigma0=None):
         super(SMC, self).__init__(log_posterior, x0, sigma0)
 
-        # Number of particles
-        self._particles = 1000
-
-        # Temperature schedule
-        self._schedule = None
-        self.set_temperature_schedule()
-
-        # ESS threshold (default from Del Moral et al.)
-        self._ess_threshold = self._particles / 2
-
-        # Determines whether to resample particles at end of
-        # steps 2 and 3 from Del Moral et al. (2006)
-        self._resample_end_2_3 = True
-
-        # Set number of MCMC steps to do for each distribution
-        self._kernel_samples = 1
-
-    def set_particles(self, particles):
-        """
-        Sets the number of particles
-        """
-        if particles < 10:
-            raise ValueError('Must have more than 10 particles in SMC.')
-        self._particles = particles
-
-    def set_resample_end_2_3(self, resample_end_2_3):
-        """
-        Determines whether a resampling step is performed at end of
-        steps 2 and 3 in Del Moral et al. Algorithm 3.1.1
-        """
-        if not isinstance(resample_end_2_3, bool):
-            raise ValueError('Resample_end_2_3 should be boolean.')
-        self._resample_end_2_3 = resample_end_2_3
-
-    def set_ess_threshold(self, ess_threshold):
-        """
-        Sets the threshold ESS (effective sample size)
-        """
-        if ess_threshold > self._particles:
-            raise ValueError('ESS threshold must be below actual sample size.')
-        if ess_threshold < 0:
-            raise ValueError('ESS must be greater than zero.')
-        self._ess_threshold = ess_threshold
-
-    def set_kernel_samples(self, kernel_samples):
-        """
-        Sets number of MCMC samples to do for each temperature
-        """
-        if kernel_samples < 1:
-            raise ValueError('Number of samples per temperature must be >= 1.')
-        if not isinstance(kernel_samples, int):
-            raise ValueError('Number of samples per temperature must be int.')
-        self._kernel_samples = kernel_samples
-
-    def set_temperature_schedule(self, schedule=10):
-        """
-        Sets a temperature schedule.
-
-        If ``schedule`` is an ``int`` it is interpreted as the number of
-        temperatures and a schedule is generated that is uniformly spaced on
-        the log scale.
-
-        If ``schedule`` is a list (or array) it is interpreted as a custom
-        temperature schedule.
-        """
-
-        # Check type of schedule argument
-        if np.isscalar(schedule):
-
-            # Set using int
-            schedule = int(schedule)
-            if schedule < 2:
-                raise ValueError(
-                    'A schedule must contain at least two temperatures.')
-
-            # Set a temperature schedule that is uniform on log(T)
-            a_max = 0
-            a_min = np.log(0.0001)
-            #diff = (a_max - a_min) / schedule
-            log_schedule = np.linspace(a_min, a_max, schedule)
-            self._schedule = np.exp(log_schedule)
-            self._schedule.setflags(write=False)
-
-        else:
-
-            # Set to custom schedule
-            schedule = pints.vector(schedule)
-            if len(schedule) < 2:
-                raise ValueError(
-                    'A schedule must contain at least two temperatures.')
-            if schedule[0] != 0:
-                raise ValueError(
-                    'First element of temperature schedule must be 0.')
-
-            # Check vector elements all between 0 and 1 (inclusive)
-            if np.any(schedule < 0):
-                raise ValueError('Temperatures must be non-negative.')
-            if np.any(schedule > 1):
-                raise ValueError('Temperatures cannot exceed 1.')
-
-            # Store
-            self._schedule = schedule
-
     def run(self):
         """ See :meth:`SMCSampler`. """
-        # Report the current settings
-        if self._verbose:
-            print('Running sequential Monte Carlo')
-            print('Total number of particles: ' + str(self._particles))
-            print('Number of temperatures: ' + str(len(self._schedule)))
-            if self._resample_end_2_3:
-                print('Resampling at end of each iteration')
-            else:
-                print('Not resampling at end of each iteration')
-            print(
-                'Number of MCMC steps at each temperature: '
-                + str(self._kernel_samples))
-
-        # Initial starting parameters
-        mu = self._x0
-        sigma = self._sigma0
 
         # Starting parameters
         samples = np.random.multivariate_normal(
@@ -178,45 +59,6 @@ class SMC(pints.SMCSampler):
             m_samples[:, :, i + 1] = samples_new
         self._weights = weights_new
         return m_samples[:, :, -1]
-
-    def _tempered_distribution(self, x, beta):
-        """
-        Returns the tempered log-pdf:
-        ``beta * log pi(x) + (1 - beta) * log prior(x)``
-        If not explicitly given prior is assumed to be multivariate normal.
-        """
-        return beta * self._log_posterior(x) + (1 - beta) * self._log_prior(x)
-
-    def _w_tilde(self, x_old, x_new, beta_old, beta_new):
-        """
-        Calculates the log unnormalised incremental weight as per eq. (31) in
-        Del Moral.
-        """
-        return (
-            self._tempered_distribution(x_old, beta_new)
-            - self._tempered_distribution(x_old, beta_old)
-        )
-
-    def _new_weight(self, w_old, x_old, x_new, beta_old, beta_new):
-        """
-        Calculates the log new weights as per algorithm 3.1.1. in Del Moral et
-        al. (2006).
-        """
-        w_tilde_value = self._w_tilde(x_old, x_new, beta_old, beta_new)
-        return np.log(w_old) + w_tilde_value
-
-    def _new_weights(
-            self, w_old, samples_old, samples_new, beta_old, beta_new):
-        """
-        Calculates the new weights as per algorithm 3.1.1 in Del Moral et al.
-        (2006).
-        """
-        w_new = np.array([
-            self._new_weight(
-                w, samples_old[i], samples_new[i], beta_old, beta_new)
-            for i, w in enumerate(w_old)])
-
-        return np.exp(w_new - logsumexp(w_new))
 
     def _kernel_sample(self, samples, beta):
         """
